create photo selector component for chat dialog & user profile form
